libc: fix build in C23 mode
GCC 15 defaults to C23, in which bool, true, and false are keywords.

Signed-off-by: Yaakov Selkowitz <yselkowi@redhat.com>
(cherry picked from commit 556bc78b89338f9bead554aa5dfb3a65d60b4fba)
References: bsc#1241473

For now, take this patch to unblock the build. However, upstream is
still discussing on what the best solution actually is, so it
might be that this will be reverted and another one applied.

Signed-off-by: Dario Faggioli <dfaggioli@suse.com>